requirements.txt
# Core libraries
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0

# API integrations
alpaca-py==0.21.0          # For paper/live trading
anthropic==0.18.1          # For Claude API
openai==1.12.0             # For OpenAI API (if using)
groq==0.4.2                # For Groq API

# Data and analysis
yfinance==0.2.36           # For market data
ta-lib==0.4.28             # Technical analysis (optional)
#pandas-ta==0.3.14b0         # Or latest: check pypi.org/project/pandas-ta
scipy==1.11.4              # Statistical functions
statsmodels==0.14.1        # Statistical models

# Database
psycopg2-binary==2.9.9     # PostgreSQL (if using)
sqlalchemy==2.0.25         # ORM

# Utilities
requests==2.31.0
aiohttp==3.9.1             # Async HTTP
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Monitoring
prometheus-client==0.19.0  # Metrics (optional)
